+++ src/main/java/me/itzsomebody/radon/transformers/linenumbers/RemoveLineNumbers.java
@@ -43,7 +43,7 @@ public class RemoveLineNumbers extends AbstractTransformer {
                 !this.exempted(classNode.name, "LineNumbers")).forEach(classNode -> {
             classNode.methods.stream().filter(methodNode ->
                     !this.exempted(classNode.name + '.' + methodNode.name + methodNode.desc, "LineNumbers")
-                            && !Modifier.isAbstract(methodNode.access)).forEach(methodNode -> {
+                            && hasInstructions(methodNode)).forEach(methodNode -> {
                 for (AbstractInsnNode insn : methodNode.instructions.toArray()) {
                     if (insn instanceof LineNumberNode) {
                         methodNode.instructions.remove(insn);