+++ src/main/java/me/itzsomebody/radon/transformers/stringencryption/SuperLightStringEncryption.java
@@ -63,7 +63,7 @@ public class SuperLightStringEncryption extends AbstractTransformer {
         this.classNodes().stream().filter(classNode -> !this.exempted(classNode.name, "StringEncryption")).forEach(classNode -> {
             classNode.methods.stream().filter(methodNode ->
                     !this.exempted(classNode.name + '.' + methodNode.name + methodNode.desc, "StringEncryption")
-                            && !BytecodeUtils.isAbstractMethod(methodNode.access)).forEach(methodNode -> {
+                            && hasInstructions(methodNode)).forEach(methodNode -> {
                 for (AbstractInsnNode insn : methodNode.instructions.toArray()) {
                     if (methodSize(methodNode) > 60000) break;
                     if (insn instanceof LdcInsnNode) {